refactor: move `idt` into `interrupt`
